Refactor simple pizza factory
