+++ src/main/java/org/mclods/simple_factory/product/impl/VeggiePizza.java
@@ -1,6 +1,6 @@
-package org.mclods.simple_factory.Pizza.impl;
+package org.mclods.simple_factory.product.impl;
 
-import org.mclods.simple_factory.Pizza.Pizza;
+import org.mclods.simple_factory.product.Pizza;
 
 public class VeggiePizza extends Pizza {
     public VeggiePizza() {
